fix SQL binding error on admin dashboard for refs to null components within events
Previous version had a 1:1 relationship between the event and the tool.

We can now have events that can have no particular tool ... so the
relationship is now 1:(0 or 1)

This means that any SQL that gathers event data  may return NULL data in
columns that include the tool that is referenced from the event.